+++ src/app/home/home.component.css
@@ -89,7 +89,6 @@
     font-size: 10px;
     font-weight: bold;
     font-variant: small-caps;
-    display: inline-block;
     float: right;
     margin-top: 0px;
 }
